+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         android:theme="@style/Theme.DoNotLate"
         tools:targetApi="31">
         <activity
-            android:name=".feature.chatroom.presentation.view.ChatRoomActivity"
+            android:name=".feature.mypromise.presentation.view.MyPromiseAdapter"
             android:exported="false" />
         <activity
             android:name=".feature.friends.presentation.view.FriendsActivity"
