--- app/src/main/java/com/example/donotlate/feature/chatroom/presentation/adapter/ChatRoomAdapter.kt
@@ -1,4 +1,0 @@
-package com.example.donotlate.feature.chatroom.presentation.adapter
-
-class ChatRoomAdapter {
-}
