+++ app/src/main/java/com/example/donotlate/feature/main/presentation/view/MainFragment.kt
@@ -18,7 +18,7 @@ import com.example.donotlate.feature.friends.presentation.view.FriendsActivity
 import com.example.donotlate.feature.main.presentation.viewmodel.MainPageViewModel
 import com.example.donotlate.feature.main.presentation.viewmodel.MainPageViewModelFactory
 import com.example.donotlate.feature.minigame.MiniGameFragment
-import com.example.donotlate.feature.mypromise.MypromiseListFragment
+import com.example.donotlate.feature.mypromise.presentation.view.MypromiseListFragment
 import com.example.donotlate.feature.room.presentation.view.RoomActivity
 import com.example.donotlate.feature.searchPlace.presentation.search.PlaceSearchFragment
 import com.example.donotlate.feature.setting.presentation.view.SettingFragment
