+++ app/src/main/java/com/example/donotlate/feature/mypromise/presentation/adapter/MyPromiseAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.donotlate.feature.mypromise.presentation.adapter
+
+class MyPromiseAdapter {
+}
