+++ app/src/main/java/com/example/donotlate/feature/mypromise/presentation/viewmodel/MyPromiseViewModel.kt
@@ -1,21 +1,9 @@
-package com.example.donotlate.feature.chatroom.presentation.viewmodel
+package com.example.donotlate.feature.mypromise.presentation.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.example.donotlate.core.domain.usecase.AcceptFriendRequestsUseCase
-import com.example.donotlate.core.domain.usecase.GetCurrentUserUseCase
-import com.example.donotlate.core.domain.usecase.GetFriendRequestsListUseCase
-import com.example.donotlate.core.domain.usecase.GetFriendRequestsStatusUseCase
-import com.example.donotlate.core.domain.usecase.GetFriendsListFromFirebaseUseCase
-import com.example.donotlate.core.domain.usecase.GetUserDataUseCase
-import com.example.donotlate.core.domain.usecase.MakeAFriendRequestUseCase
-import com.example.donotlate.core.domain.usecase.SearchUserByIdUseCase
-import com.example.donotlate.feature.chatroom.domain.usecase.MakeAPromiseRoomUseCase
-import com.example.donotlate.feature.friends.presentation.mapper.toModel
-import com.example.donotlate.feature.friends.presentation.model.FriendsUserModel
-import com.example.donotlate.feature.friends.presentation.viewmodel.FriendsViewModel
+import com.example.donotlate.feature.room.domain.usecase.MakeAPromiseRoomUseCase
 import com.example.donotlate.feature.main.presentation.model.UserModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
