+++ app/src/main/java/com/example/donotlate/feature/room/presentation/dialog/ResultFragmentDialog.kt
@@ -8,11 +8,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.Window
 import androidx.fragment.app.DialogFragment
-import com.example.donotlate.MainActivity
 import com.example.donotlate.R
 import com.example.donotlate.databinding.BackDialogBinding
-import com.example.donotlate.feature.chatroom.presentation.view.ChatRoomActivity
-import com.example.donotlate.feature.chatroom.presentation.view.ChatRoomFragment
 import com.example.donotlate.feature.room.presentation.view.RoomActivity
 import com.example.donotlate.feature.room.presentation.view.RoomResultFragment
