+++ app/src/main/res/layout/fragment_chat_room.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.chatroom.presentation.view.ChatRoomFragment">
+    tools:context=".feature.mypromise.presentation.view.MyPromiseRoomFragment">
 
     <ImageView
         android:id="@+id/iv_chat_room_back"
