+++ app/src/main/res/layout/fragment_chat_room_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.chatroom.presentation.view.ChatRoomListFragment">
+    tools:context=".feature.mypromise.presentation.view.ChatRoomListFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
