+++ app/src/main/res/layout/fragment_mypromise_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.mypromise.MypromiseListFragment">
+    tools:context=".feature.mypromise.presentation.view.MypromiseListFragment">
 
 
     <ImageView
